refactor: 테스트명 Repository -> DAO 변경
구현체가 여러 개 일수도 있고 DAO에 대한 테스트라는 것을 명확화
