+++ src/test/java/roomescape/config/TestConfig.java
@@ -5,11 +5,14 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.Collections;
 import javax.sql.DataSource;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 
+@TestConfiguration
 public class TestConfig {
 
     public static DataSource getDataSource() {
@@ -45,5 +48,10 @@ public class TestConfig {
         } catch (IOException e) {
             throw new RuntimeException("파일을 가져오는 데 문제가 생겼습니다!");
         }
+    }
+
+    @Bean
+    public JdbcTemplate jdbcTemplate() {
+        return getJdbcTemplate();
     }
 }
